+++ R/lme4_utilities.R
@@ -153,7 +153,7 @@ extractVarCov.merMod <- function(m){
 #' @param B the bootstrap sample size
 #' @export bootinvFIM.merMod
 #' @export
-bootinvFIM.merMod <- function(m, B=1000){
+bootinvFIM.merMod <- function(m, B=1000, seed=0){
   
   mySumm <- function(m,diagSigma=F) {
     beta <- lme4::fixef(m)
@@ -173,6 +173,10 @@ bootinvFIM.merMod <- function(m, B=1000){
   
   # Use bootMer functions if linear or generalized linear, otherwise code our own bootstrap
   if (!nonlin){
+    
+    message(paste0("\t ...generating the B=",B," bootstrap samples ...\n"))
+    
+    set.seed(seed)
     bootstrap <- lme4::bootMer(m, mySumm, use.u = F, type = "parametric", nsim = B)
     bootstrap <- bootstrap$t[, colSums(bootstrap$t != 0) > 0]
     invfim <- cov(bootstrap)
@@ -205,6 +209,7 @@ bootinvFIM.merMod <- function(m, B=1000){
     tbar <- utils::txtProgressBar(min=1,max=B,char = ".", style = 3)
     grpVar <- m@frame[,grpFactor]
     nmeInd <- unique(grpVar)
+    set.seed(seed)
     while (b <= B){  
       utils::setTxtProgressBar(tbar,b)      
       phi <- t(t(chol(Sigma))%*%matrix(stats::rnorm(nrandEfft*nind,0,1),ncol=nind))